requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
transformers>=4.30.0
python-dotenv>=0.21.0
google-genai>=0.2.0

